Add files via upload
Hi everyone,

This is a dump from the server. It'll be missing comments, logic, standardisation, security, etc etc etc.

For those wondering why the database / php vs json:

This originally started as just the lights and timers, using MariaDB on a Raspberry Pi Zero W. 

simpleLifter was an addon after I realised MS Excel was too limited. But for some reason decided to use JSON instead of continuing with the DB... I dunno...

I only have the DB & server integrated simpleLifter version up. There was also a local version which is good for a computer with 2 or 3 screen outputs and can drive the platform display and warmup area all locally. Right now, very few latency or availability issues with the server so I've removed it.

TODO:
* A lot...